etc: 북노트 스텝업 컴포넌트 분리
