+++ src/components/bookNote/stepUp/LeftWrapper.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import styled from "styled-components";
+
+export default function LeftWrapper() {
+  return (
+    <StLeftWrapper>
+      <StImgWrapper>
+        <img src="" alt="" />
+      </StImgWrapper>
+      <StLifeQuotes></StLifeQuotes>
+    </StLeftWrapper>
+  );
+}
+const StLeftWrapper = styled.div``;
+const StImgWrapper = styled.div``;
+const StLifeQuotes = styled.p``;
